backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn==0.27.0

# Database
sqlalchemy==2.0.25

# Data validation
pydantic==2.5.3

# File uploads
python-multipart==0.0.6

# HTTP client
httpx==0.26.0

# Vector database
chromadb==0.5.0

# Data processing
pandas==2.1.4

# Environment variables
python-dotenv==1.0.0

# Async file operations
aiofiles==23.2.1

# Google ADK for multi-agent system
google-adk>=0.3.0
google-genai>=1.0.0
